model updates

+++ src/structs/point.rs
@@ -1,18 +1,17 @@
-use std::f32::NAN;
 
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Clone, Debug)]
 pub struct Point {
-    pub open: f32,
-    pub high: f32,
-    pub low: f32,
-    pub close: f32,
+    pub open: f64,
+    pub high: f64,
+    pub low: f64,
+    pub close: f64,
     pub volume: u32,
 }
 
 impl Point {
-    pub fn new(open: f32, high: f32, low: f32, close: f32, volume: u32) -> Self {
+    pub fn new(open: f64, high: f64, low: f64, close: f64, volume: u32) -> Self {
         Self {
             open,
             high,
